Inline site footer styles with tailwind classes
Now that we've moved this to a partial, no need to keep the styles as a component class. We did that for others to be able to easily override using CSS but no point in that since the footer is really meant for the user to completely customize without any defaults applied.