Revert nested submenus from #5994
We've already removed SASS so this is a revert of the behavioral change. It's not recommend for main navigations to have more than a single nested menu level. Further since the menu only rendered horizontally when the change was introduced, we now render the menu vertically as a drawer style component for mobile support.